Crate address_book

Address Book

The Espresso Address Book maps user addresses to encryption public keys. Users publish their encryption public keys here. This is done by submitting their UserPubKey bundle (that includes both address and encryption public key) and signing the request using the user private key associated with the address. This way, it is infeasible to publish an invalid encryption key for a user.

This public mapping allows a sender to retrieve the encryption key associated with the recipient address. This is particularly important for the freezing feature: when the freezer releases records back to the original owners, it needs to know their encryption keys in order to produce owner memos for them.

The Address Book is a standalone web service.

The port defaults to 50078, but can be overridden at startup by setting the environment variable ESPRESSO_ADDRESS_BOOK_PORT.

POST insert_pubkey - Insert or update the public key at the given address.

POST request_pubkey - Fetch the public key for the given address. If not found, return StatusCode::NotFound.

See tests/tests.rs for usage examples.

Modules

error

signal

store

Structs

Args

Command line arguments

InsertPubKey

Structure to supply parameters to web requests

ServerState

Web server state

Enums

AppKey

Lookup keys for application-specific configuration settings

Constants

APP_NAME

Application name used for locating configuration files

Functions

eq_header_values

init_web_server

Initialize the web server

insert_pubkey

verify_sig_and_get_pub_key

Lookup a user public key from a signed public key address.

Type Definitions

Result